UI, Composure, Initialization fixes/tweaks
Fixes drag/unlick event errors when clicked/dragging elements are destroyed before the end of said event.
Fixes the composure removal logic to detect the removal at an accelerated rate.
Changes (removes) the initialization logic, it seemed unecessary, leaving the mid-cutscene alert in place. (UI won't show till end of cutscene, so that's more relevant of an alert)
Added a delay to sms_initialize on job change to give windower's memory a chance to update itself, as it was still reporting buffs that had worn on job change.