Keep Hover Synced-but-blocked on Click-n-drag-away
Keeps the hover variable firing on dragaway - but with a new event variable that allow the listener functions to ignore embedded hovers - so that the on-release hover event has an updated hover.t. Previously, on unclick on a different-but-valid element, said element wouldn't get hovered since it didn't belong to the hover variable since we weren't keeping it updated on the embedded hovers. Now it stays synced.